+++ weldx/__init__.py
@@ -2,7 +2,6 @@ import warnings
 
 # asdf extensions and tags
 import weldx.asdf
-
 # geometry packages
 import weldx.geometry
 import weldx.transformations
@@ -19,4 +18,4 @@ with warnings.catch_warnings():
 __version__ = get_versions()["version"]
 del get_versions
 
-__all__ = ["geometry", "transformations", "utility", "asdf", "Q_"]
+__all__ = ["geometry", "measurement", "transformations", "utility", "asdf", "Q_"]