+++ weldx/asdf/tags/weldx/__init__.py
@@ -1,1 +1,1 @@
-from . import aws, core, debug, time, unit
+from . import aws, core, debug, equipment, measurement, time, unit