+++ weldx/asdf/tags/weldx/measurement/measurement_chain.py
@@ -1,0 +1,26 @@
+from weldx.asdf.types import WeldxType
+from weldx.asdf.utils import drop_none_attr
+from weldx.measurement import MeasurementChain
+
+__all__ = ["MeasurementChain", "MeasurementChainType"]
+
+
+class MeasurementChainType(WeldxType):
+    """Serialization class for measurement chains"""
+
+    name = "measurement/measurement_chain"
+    version = "1.0.0"
+    types = [MeasurementChain]
+    requires = ["weldx"]
+    handle_dynamic_subclasses = True
+
+    @classmethod
+    def to_tree(cls, node: MeasurementChain, ctx):
+        """convert to tagged tree and remove all None entries from node dictionary"""
+        tree = drop_none_attr(node)
+        return tree
+
+    @classmethod
+    def from_tree(cls, tree, ctx):
+        obj = MeasurementChain(**tree)
+        return obj