+++ weldx/transformations.py
@@ -9,6 +9,7 @@ from typing import Hashable, List, Union
 import networkx as nx
 import numpy as np
 import pandas as pd
+import pint
 import xarray as xr
 from scipy.spatial.transform import Rotation as Rot
 
@@ -373,7 +374,7 @@ class LocalCoordinateSystem:
                 pass
 
             if not isinstance(coordinates, xr.DataArray):
-                if not isinstance(coordinates, np.ndarray):
+                if not isinstance(coordinates, (np.ndarray, pint.Quantity)):
                     coordinates = np.array(coordinates)
                 time_coordinates = None
                 if coordinates.ndim == 2: